requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
bangladesh==0.0.10
constantly==15.1.0
Django==4.1.1
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==21.3.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tornado==6.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2022.4
zipp==3.8.1
zope.interface==5.4.0
